libxl: add support for channels
A 'channel' is a low-bandwidth private communication channel that
resembles a physical serial port.

This patch adds a list of channels to the IDL for the domain config.
Each channel has a 'kind' which describes what should happen to
the data. Currently defined kinds are:

  * PTY: the I/O surfaces as a pty in the backend domain
  * SOCKET: a listening Unix domain socket accepts a connection in
    the backend domain and proxies

Channels are implemented using PV consoles backed by qemu (not
xenconsoled).

Channels may be listed but don't currently support hotplug/unplug.

Signed-off-by: David Scott <dave.scott@citrix.com>