xl: add support for channels
This adds support for channel declarations of the form:

  channel = [ "name=...,kind=...[,path=...][,backend=...]" ]

where 'name' is a label to identify the channel to the frontend.

If 'kind = pty' then the channel is connected to a pty in the
  backend domain
If 'kind = socket' then the channel is connected to a Unix domain
  socket given by 'path = ...' in the backend domain.

This patch also adds the command:

  xl channel-list <domain>

which allows the state of channels to be queried. In particular if
'kind=pty' this will show the path of the pty slave device.

Signed-off-by: David Scott <dave.scott@citrix.com>